+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All Notable changes to `ElementFinder` will be documented in this file
 
 ### Removed
 
+### Added
+ - #50 Add `StringCollection::unique` function
+  
 ## 0.1.0-alpha.5 [2017-03-10]
 
 ### Added